==> 01_data cleaning.Rmd <==
---
title: "Data Cleaning"
author: "Zhaoqi Liu"
date: "10/13/2020"
output: pdf_document
---
####input data (per patient)
```{r}
library(readxl)
LSRdf <- read_excel("Composite TMA data V4.xlsx", sheet = "LSR data")
LSRperson<- na.omit(LSRdf)
LSRperson1<-LSRperson

Clinicaldf<-read_excel("Composite TMA data V4.xlsx", sheet = "AllClinicalInfo")
Clinicaldf[,10]<-as.numeric(unlist(Clinicaldf[,10])) #convert types into number from character 
Clinicaldf[,11]<-as.numeric(unlist(Clinicaldf[,11]))
```

```{r}
#change the id number by deleting the first digit of each id, for data merging 
id<-LSRperson1[,1]
ID<-substring(sapply(id,as.character),2)
LSRperson$ID<-sapply(ID,as.numeric)
LSRperson #348 patients of LSR data, no duplicate
```

```{r}
#select rows that the survival month is not NA
Cperson2<-Clinicaldf[!is.na(Clinicaldf[,19]),] 
Cperson1<-Cperson2
id<-Cperson2[,1]
ID<-substring(sapply(id,as.character),2)
Cperson1$ID<-sapply(ID,as.numeric)
#eliminate duplicate rows 
#sum(ifelse(duplicated(Cperson1)==TRUE,1,0)) #20 duplicates
Cperson<-Cperson1[!duplicated(Cperson1),]
Cperson
```

```{r}
###Merge two dataset based on ID 
library(dplyr)
#two obs on LSR data do not have a match on clinical data (ID 978, 1463)
df<-left_join(Cperson,LSRperson,by = "ID")
df
#treat some variables as factor? 
```


```{r}
#the dataframe before variable selection 346 observations and 26 variables
saveRDS(df,"breastcancerdf.rds")
```

```{r}
df<-readRDS("breastcancerdf.rds")
```


Select: 
- 8 variables that are used to draw the Kaplan Meier Plots:
race(DEMO_RACE),age(DEMO_AGE_AT_DIAGNOSIS), ER(PATH_ER), PR(PATH_PR), HER2(PATH_HER2), grade(PATH_SURGERYOVERALLGRADE), stage, menopause(HORMO_HORMO_MENOPAUSESTATUS)  
- 5 death and relapse status variables (responses) 
- 3 variables that measure the average LSR counts per patients
Then, we change the variable names for convenience. 
```{r}
bcdf<-df[,c(1:3,6:8,13,20,4,15:16,19,18,17,24:26)]
names(bcdf)<-c("Id","Race","Age","ER","PR","HER2","Grade","Stage","Menopause",
                "Metastatic","Survival","Survival_month","Relapse","Relapse_month",
               "AveNCount", "AveCMCount", "AveTCount")
bcdf #dataframe after selecting the necessary variables, 346 observations and 17 variables

```

United the relapse group and died group of patients. 
Because the relapse = 1 only means measured relapse. It's possible that the patient died of cancer recurrence without a diagnosis. Thus, we add two new columns: Mod_Relapse and Mod_Relapse_month. The Mod_Relapse variable denotes the status of measured relapse or death of disease. The status are coded as 1 if the patient either had a relapse or died. (Relapse=1 or Survival=1), 0 otherwise. There are 111 patients with Mod_Relapse denoted as 1. The Mod_Relapse_month refers to the time from breast cancer diagnosis to diagnosed recurrence or death, whichever comes first. If the patient with Mod_Relapse denoted as 0, their Mod_Relapse_month value is NA. 
```{r}
bcdf$Mod_Relapse<-ifelse(bcdf$Relapse==1 | bcdf$Survival ==1,1,0)
bcdf$Mod_Relapse_month<-ifelse(bcdf$Relapse==0 & bcdf$Survival==1, bcdf$Survival_month, bcdf$Relapse_month)
bcdf<-bcdf[,c(1:14,18,19,15:17)]#change the order of column 
head(bcdf) #17 variables 
```


Delete observations that race are specified as "other". We only focus on "Black" and "White" in race.  
Removing 18 observations whose Grade value is missing
```{r}
bcdf<-bcdf[bcdf$Race!="Other",] # 341 observations, 17 variables
bcdf<-bcdf[!is.na(bcdf$Grade),] 
summary(bcdf[,-1])

```

Convert most clinical info variables to factors (race, ER, PR, HER2, grade, stage, menopause, metastatic). Not convert survival and relapse because survival analyses require them to be numeric events.

```{r}
bcdf[,-c(1,3,11:19)]<-lapply(bcdf[,-c(1,3,11:19)],as.factor)
summary(bcdf)
```





Save the resulting dataset 
- dataset contains 341 observations and 17 variables.
```{r}
saveRDS(bcdf,"bcdf_final.rds")
```
