==> 02_Kaplan Meier Plots_ver1.Rmd <==
---
title: "Kaplan-Meier plots"
author: "Zhaoqi Liu"
date: "2/20/2021"
output: pdf_document
---

```{r}
bcdf<-readRDS("breastcancerdf.rds")
```

```{r}
library(dplyr)
library(survival)
library(survminer)
```

###Pre-select
Select 8 variables that are used to draw the Kaplan Meier Plots:
race(DEMO_RACE),age(DEMO_AGE_AT_DIAGNOSIS), ER(PATH_ER), PR(PATH_PR), HER2(PATH_HER2), grade(PATH_SURGERYOVERALLGRADE), stage, menopause(HORMO_HORMO_MENOPAUSESTATUS)  
and 5 death and relapse status variables (response)
Then, we change the variable names for convenience. 


```{r}
kmpdf<-bcdf[,c(1:3,6:8,13,20,4,15:16,19,18,17)]
names(kmpdf)<-c("id","race","age","ER","PR","HER2","grade","stage","menopause",
                "metastatic","survival","survival_month","relapse","relapse_month")
```

```{r}
head(kmpdf)
```

```{r}
median(kmpdf$age)
mean(kmpdf$age)
hist(kmpdf$age,freq=FALSE,main="histogram of age", xlab="age" )
lines(density(kmpdf$age),lwd=2, col=2) #kernel density plot 

```

The density plot shows that the age variable is approximately normal distributed. Since mean(55.58671) is slightly greater than median(54), the variable is slightly right skewed. Divide age into two groups by the median of age. If age is below the 54, we note it as "young"; otherwise "old".  
```{r}
kmpdf$age<-ifelse(kmpdf$age<median(kmpdf$age),"young","old")
head(kmpdf)
```

Delete observations that race are specified as "other". We only focus on "Black" and "White" in race.  
```{r}
kmpdf<-kmpdf[kmpdf$race!="Other",]
summary(kmpdf[,-1])
```


Convert most variables to factors except id and the response(survial, survival months, relapse, and relapse month).  Not convert survival and relapse because survival analyses require them to be numeric events.

```{r}
kmpdf[,-c(1,11:14)]<-lapply(kmpdf[,-c(1,11:14)],as.factor)
summary(kmpdf)
sapply(c(11,13),function(x){table(kmpdf[,x])})
```

```{r}
saveRDS(kmpdf,"kmplotdf.rds")
```

```{r}
kmpdf<-readRDS("kmplotdf.rds")
```


###Plot the density of survival month and relapse month 
survival: 0: alive 1:dead
Survival Month: Number of months survived since breast cancer diagnosis, value calculated automatically from DEMO_DIAGNOSISDATE to FOLLOW_LAST_CONTACT_DATE or FOLLOW_DEATHDATE

relapse: 0: no relapse 1: local and/or distant cancer recurrence or died of disease
Relapse event status (0=no relapse, 1=local and/or distant cancer recurrence or died of disease) NOT Relapse-Free Survival
Relapse Month: Number of months to relapse  (number of months that is free from any relapse: local/distant cancer recurrence, or dead), calculated automatically from DEMO_DIAGNOSISDATE to FOLLOW_LOCALRECURRENCEDATE, and/or FOLLOW_DISTANT_DISEASE_DX_DATE. NOT Relapse-Free Survival

```{r}
table(survival,relapse)
```

230 patients survived and didn't suffer a relapse. 14 people had relapsed but was still alive at the time of follow-up. 46 patients died without a relapse, and 51 patients died and suffered a relapse.  

```{r}
hist(kmpdf$survival_month,breaks =seq(0,250,25) ,freq=FALSE,xlim=c(0,250),
     main="Histogram of Survival Months",xlab= "survival months")
lines(density(kmpdf$survival_month),lwd=2,col=2)
```

```{r}
hist(kmpdf$relapse_month,breaks =seq(0,175,25) ,freq=FALSE,xlim=c(0,175),
     main="Histogram of Relapse Months",xlab= "relapse months")
lines(density(kmpdf$relapse_month,na.rm = TRUE),lwd=2,col=2)
```

The distribution of survival months is quite normal, even though it's left skewed a little bit, which means that more observations than expected have short survival months. There are only 65 observations that has the relapse months data. The histogram shows that the data maybe follow a poission distribution with a small parameter $\lambda$. 


```{r}
attach(kmpdf)
```

### Kaplan Meier Curves

####Overall
Compare the survival distribution to examine whether or not there is an association between features and length of survival
```{r}
surv.all<-survfit(Surv(survival_month,survival)~1)
summary(surv.all)
plot(surv.all,main="Plot of Survival Curve for Breast Cancer Patients",xlab= "Length of Survival", ylab= "Proportion of Individuals who have Survived")
```

For the 341 people in the dataset(keeps NA in the grade), 97 people were uncensored(followed for the entire time, until occurence of event). Since the data has not yet dropped to 50% survival at the end of the available data, there is an NA value for median survival.
The following summary goes through each time point in the study in which an individual was lost to follow up or died and re-computes the total number of people still at risk (n.risk), the number of events at that time point (n.event), the proportion of individuals who survived up until that point (survival) and the standard error (std.err) and 95% confidence interval (lower 95% CI, upper 95% CI) for the proportion of individuals who survived at that point.
This plot shows the survival curve (also known as a Kaplan-Meier plot), the proportion of individual who have survived up until that particular time as a solid black line and the 95% confidence interval (the dashed lines).

#### Race 
```{r}
surv.race<-survfit(Surv(survival_month,survival)~race)
plot(surv.race,col=c("blue","red"),ylim = c(0,1),
     main="Plot of Survival Curve by Race Group", 
     xlab = "Length of Survival",ylab= "Proportion of Individuals who have survived")
legend("bottomright",legend=c("Black","White"),fill=c("blue","red"),bty="n")
#Since the levels are "Black","White"
survdiff(Surv(survival_month,survival)~race)
```

The 95% confidence interval of survival time for those on maintained chemotherapy is (, NA); NA in this case means infinity. A 95% upper confidence limit of NA/infinity is common in survival analysis due to the fact that the data is skewed.

Using `survminer` package to plot. 
```{r}
ggsurvplot(surv.race,data=kmpdf,censor.size=4,conf.int=TRUE,pval=TRUE,
           legend.labs= c("Black","White"))
```

relapse:

```{r}
relapse.race<-survfit(Surv(relapse_month,relapse)~race)
plot(relapse.race,col=c("blue","red"),ylim = c(0,1),
     main="Plot of Relapse Curve by Race Group", 
     xlab = "Length of Relapse",ylab= "Proportion of Individuals who have relapsed")
legend("topright",legend=c("Black","White"),fill=c("blue","red"),bty="n")
#Since the levels are "Black","White"
survdiff(Surv(survival_month,survival)~race)
```



#### Age
```{r}
surv.age<-survfit(Surv(survival_month,survival)~age)
plot(surv.age,col=c("blue","red"),ylim = c(0,1),
     main="Plot of Survival Curve by Age Group", 
     xlab = "Length of Survival",ylab= "Proportion of Individuals who have survived")
legend("bottomright",legend=c("Old","Young"),fill=c("blue","red"),bty="n")
```

log-rank test 
H0: There is no difference in the survival function between those who were young and those who were old
```{r}
survdiff(Surv(survival_month,survival)~age)
#reject H0
```





#### ER
estrogen receptor status 
```{r}
surv.ER<-survfit(Surv(survival_month,survival)~ER)
plot(surv.ER,col=c("blue","red"),ylim = c(0,1),
     main="Plot of Survival Curve by ER status", 
     xlab = "Length of Survival",ylab= "Proportion of Individuals who have survived")
legend("bottomright",legend=c("negative","positive"),fill=c("blue","red"),bty="n")
#levels = 0,1 0 means negative, 1 means positive

survdiff(Surv(survival_month,survival)~ER) #not reject
```




#### PR
progesterone receptor status
```{r}
surv.PR<-survfit(Surv(survival_month,survival)~PR)
plot(surv.PR,col=c("blue","red"),ylim = c(0,1),
     main="Plot of Survival Curve by PR status", 
     xlab = "Length of Survival",ylab= "Survival Probability")
legend("bottomright",legend=c("negative","positive"),fill=c("blue","red"),bty="n")
#levels = 0,1 0 means negative, 1 means positive

survdiff(Surv(survival_month,survival)~PR)
```

#### HER2 
(human epidermal growth factor receptor 2) status
positive means: When a breast cell has abnormally high levels of the HER2 gene or the HER2 protein, it is called `HER2- positive`. Most patients with metastatic breast cancer have HER2-negative breast cancer.
```{r}
surv.HER2<-survfit(Surv(survival_month,survival)~HER2)
plot(surv.HER2,col=c("blue","red"),ylim = c(0,1),
     main="Plot of Survival Curve by HER2 status", 
     xlab = "Length of Survival",ylab= "Survival Probability")
legend("bottomright",legend=c("negative","positive"),fill=c("blue","red"),bty="n")
#levels = 0,1. 0 means negative, 1 means positive

survdiff(Surv(survival_month,survival)~HER2)
```



#### grade 
The overall grade of the tumor specimen at definitive surgery
0 means low grade, 1 means high grade. 
Low-grade cancer cells (also known as well-differentiated cancer cells) look more like normal cells and tend to grow and spread more slowly than high-grade cancer cells(poorly differentiated or undifferentiated cancer cells.).
```{r}
surv.grade<-survfit(Surv(survival_month,survival)~grade)
plot(surv.grade,col=c("blue","red"),ylim = c(0,1),
     main="Plot of Survival Curve by Grade Group", 
     xlab = "Length of Survival",ylab= "Survival Probability")
legend("bottomright",legend=c("low-grade","high-grade"),fill=c("blue","red"),bty="n")

survdiff(Surv(survival_month,survival)~grade)
```



#### stage
0 means low stage, 1 means high stage.
Lower stage cancers are typically less aggressive and have a better prognosis. 
Breast cancer stage is usually expressed as a number on a scale of 0 through IV — with stage 0 describing non-invasive cancers that remain within their original location and stage IV describing invasive cancers that have spread outside the breast to other parts of the body.
```{r}
surv.stage<-survfit(Surv(survival_month,survival)~stage)
plot(surv.stage,col=c("blue","red"),ylim = c(0,1),
     main="Plot of Survival Curve by Stage Group", 
     xlab = "Length of Survival",ylab= "Survival Probability")
legend("bottomright",legend=c("low-stage","high-stage"),fill=c("blue","red"),bty="n")

survdiff(Surv(survival_month,survival)~stage)
```



#### menopause
subject's menopausal status at diagnosis
0 means post-menopause, 1 means pre-menopause
```{r}
surv.menopause<-survfit(Surv(survival_month,survival)~menopause)
plot(surv.menopause,col=c("blue","red"),ylim = c(0,1),
     main="Plot of Survival Curve by Menopause Status", 
     xlab = "Length of Survival",ylab= "Survival Probability")
legend("bottomright",legend=c("post-menopause","pre-menopause"),fill=c("blue","red"),bty="n")

survdiff(Surv(survival_month,survival)~menopause)
```


This plot looks quite similar to the plot of survival curve by age, Since all old patients are most-menopause and most of young patients are pre-menopause. However, the p-value is not signifiant. 

#### metastatic
Has the subject been diagnosed with metastatic/distant disease?

```{r}
surv.metastatic<-survfit(Surv(survival_month,survival)~metastatic)
plot(surv.metastatic,col=c("blue","red"),ylim = c(0,1),
     main="Plot of Survival Curve by Metastatic Status", 
     xlab = "Length of Survival",ylab= "Survival Probability")
legend("bottomright",legend=c("No","Yes"),fill=c("blue","red"),bty="n")

survdiff(Surv(survival_month,survival)~metastatic)
table(grade,metastatic)
table(survival,metastatic)
```

The metastatic feature is dominant in survival. There are 51 patients who were diagonised with metastatic/distant disease and only 4 of them survived at the censored time. 

Include age, PR, stage, in the report results section.
